requirements.txt
asgiref==3.4.1
black==21.9b0
Django==3.2.7
djangorestframework==3.12.4
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
coverage==5.5
drf-yasg==1.20.0
rfc3339==6.2
pandas==1.3.3